fix/ref: create IImageRepo to fix image deletion

+++ TravelBooking.Application/Images/DeleteImage/DeleteImageCommandHandler.cs
@@ -1,10 +1,6 @@
 ﻿using MediatR;
-using Microsoft.AspNetCore.Http;
 using TravelBooking.Application.Common.Interfaces;
-using TravelBooking.Application.Images.UploadImage;
 using TravelBooking.Domain.Common;
-using TravelBooking.Domain.Common.Interfaces;
-using TravelBooking.Domain.Images.Entities;
 using TravelBooking.Domain.Images.Errors;
 
 namespace TravelBooking.Application.Images.DeleteImage;
@@ -12,9 +8,9 @@ namespace TravelBooking.Application.Images.DeleteImage;
 public class DeleteImageCommandHandler:IRequestHandler<DeleteImageCommand, Result>
 {
     private readonly IImageService _imageService;
-    private readonly IRepository<Image> _imageRepository;
+    private readonly IImageRepository _imageRepository;
 
-    public DeleteImageCommandHandler(IImageService imageService, IRepository<Image> imageRepository)
+    public DeleteImageCommandHandler(IImageService imageService, IImageRepository imageRepository)
     {
         _imageService = imageService;
         _imageRepository = imageRepository;
@@ -25,6 +21,10 @@ public class DeleteImageCommandHandler:IRequestHandler<DeleteImageCommand, Resul
         var image = await _imageRepository.GetByIdAsync(request.ImageId,cancellationToken);
         if(image == null) return Result.Failure(ImageErrors.ImageNotFound());
 
+        var imageUsageInfo = await _imageRepository.IsImageUsedAsThumbnailsAsync(request.ImageId);
+        if(imageUsageInfo!=null) 
+            return Result.Failure(ImageErrors.ImageInUse(imageUsageInfo.EntityType, imageUsageInfo.EntityId));
+        
         var result = await _imageService.DeleteImageAsync(image.Url);
         if(!result) return Result.Failure(ImageErrors.ErrorWhileDeleting()); 
         _imageRepository.Delete(image);
